Extract git-repo-form
