--- packages/insomnia/src/ui/components/project/hooks.tsx
@@ -1,6 +1,0 @@
-import { useState } from 'react';
-
-export function useActiveView() {
-  const [activeView, setActiveView] = useState<'project' | 'git-results'>('project');
-  return { activeView, setActiveView };
-}
